+++ app/src/test/java/com/example/shopify/feature/auth/login/LoginViewModelTest.kt
@@ -1,13 +1,10 @@
 package com.example.shopify.feature.auth.login
 
-import android.content.pm.SigningInfo
-import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.example.shopify.data.repository.shopify.FakeShopifyRepository
 import com.example.shopify.feature.auth.screens.common.AuthUIEvent
 import com.example.shopify.feature.auth.screens.login.model.SignInUserInfo
 import com.example.shopify.feature.auth.screens.login.viewModel.LoginViewModel
-import com.example.shopify.feature.auth.screens.registration.model.SignUpUserInfo
+import com.example.shopify.feature.navigation_bar.model.repository.shopify.FakeShopifyRepositoryImpl
 import com.example.shopify.feature.navigation_bar.model.repository.shopify.ShopifyRepository
 import com.example.shopify.helpers.validator.auth.UserInputValidator
 import com.example.shopify.helpers.validator.auth.UserInputValidatorImpl
@@ -18,7 +15,6 @@ import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
 import org.junit.Assert
 import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -30,7 +26,7 @@ class LoginViewModelTest {
 
     @Before
     fun init() {
-        val repository: ShopifyRepository = FakeShopifyRepository()
+        val repository: ShopifyRepository = FakeShopifyRepositoryImpl()
         val userInputValidator: UserInputValidator = UserInputValidatorImpl()
         viewModel = LoginViewModel(
             repository = repository,
