+++ app/src/test/java/com/example/shopify/feature/auth/registration/RegistrationViewModelTest.kt
@@ -2,10 +2,10 @@ package com.example.shopify.feature.auth.registration
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.example.shopify.data.repository.shopify.FakeShopifyRepository
 import com.example.shopify.feature.auth.screens.common.AuthUIEvent
 import com.example.shopify.feature.auth.screens.registration.model.SignUpUserInfo
 import com.example.shopify.feature.auth.screens.registration.viewModel.RegistrationViewModel
+import com.example.shopify.feature.navigation_bar.model.repository.shopify.FakeShopifyRepositoryImpl
 import com.example.shopify.feature.navigation_bar.model.repository.shopify.ShopifyRepository
 import com.example.shopify.helpers.validator.auth.UserInputValidator
 import com.example.shopify.helpers.validator.auth.UserInputValidatorImpl
@@ -31,7 +31,7 @@ class RegistrationViewModelTest {
 
     @Before
     fun init() {
-        val repository: ShopifyRepository = FakeShopifyRepository()
+        val repository: ShopifyRepository = FakeShopifyRepositoryImpl()
         val userInputValidator: UserInputValidator = UserInputValidatorImpl()
         viewModel = RegistrationViewModel(
             repository = repository,
