+++ app/src/test/java/com/example/shopify/feature/navigation_bar/model/repository/shopify/FakeShopifyRepositoryImpl.kt
@@ -11,7 +11,10 @@ import com.example.shopify.feature.navigation_bar.home.screen.product.model.Bran
 import com.example.shopify.feature.navigation_bar.my_account.screens.my_account.model.MinCustomerInfo
 import com.example.shopify.feature.navigation_bar.my_account.screens.order.model.order.LineItems
 import com.example.shopify.feature.navigation_bar.my_account.screens.order.model.order.Order
+import com.example.shopify.feature.navigation_bar.productDetails.screens.productDetails.model.Discount
+import com.example.shopify.feature.navigation_bar.productDetails.screens.productDetails.model.Price
 import com.example.shopify.feature.navigation_bar.productDetails.screens.productDetails.model.Product
+import com.example.shopify.feature.navigation_bar.productDetails.screens.productDetails.model.VariantItem
 import com.example.shopify.feature.navigation_bar.productDetails.screens.productDetails.view.Review
 import com.example.shopify.helpers.Resource
 import com.example.shopify.utils.Constants
@@ -25,17 +28,39 @@ import org.joda.time.chrono.ISOChronology
 
 class FakeShopifyRepositoryImpl(
 ) : ShopifyRepository {
-    private var wishList: MutableList<ID> = mutableListOf()
-    override fun signUp(userInfo: SignUpUserInfo): Flow<Resource<SignUpUserResponseInfo>> {
-        TODO("Not yet implemented")
+    private val wishList: MutableList<ID> = mutableListOf()
+
+    override fun signUp(userInfo: SignUpUserInfo): Flow<Resource<SignUpUserResponseInfo>> = flow {
+        emit(
+            Resource.Success(
+                SignUpUserResponseInfo(
+                    firstName = userInfo.firstName,
+                    lastName = userInfo.lastName,
+                    email = userInfo.email,
+                    phone = userInfo.phone,
+                    "",
+                    null
+                )
+            )
+        )
     }
 
-    override fun signIn(userInfo: SignInUserInfo): Flow<Resource<SignInUserInfoResult>> {
-        TODO("Not yet implemented")
+    override fun signIn(userInfo: SignInUserInfo): Flow<Resource<SignInUserInfoResult>> = flow {
+        emit(
+            Resource.Success(
+                SignInUserInfoResult(
+                    email = userInfo.email,
+                    password = userInfo.password,
+                    accessToken = "5",
+                    expireTime = null,
+                    error = ""
+                )
+            )
+        )
     }
 
     override suspend fun saveUserInfo(userResponseInfo: SignInUserInfoResult) {
-        TODO("Not yet implemented")
+        print(userResponseInfo)
     }
 
     override fun getUserInfo(): Flow<SignInUserInfoResult> {
@@ -78,13 +103,45 @@ class FakeShopifyRepositoryImpl(
         return Resource.Success(products)
     }
 
-    override fun getProductDetailsByID(id: String): Flow<Resource<Product>> {
-        TODO("Not yet implemented")
+    override fun getProductDetailsByID(id: String): Flow<Resource<Product>> = flow {
+        emit(
+            Resource.Success(
+                Product(
+                    images = listOf("https://www.skechers.com/dw/image/v2/BDCN_PRD/on/demandware.static/-/Sites-skechers-master/default/dw5fb9d39e/images/large/149710_MVE.jpg?sw=800"),
+                    description = "The Stan Smith owned the tennis court in the '70s." +
+                            " Today it runs the streets with the same clean," +
+                            " classic style." +
+                            " These kids' shoes preserve the iconic look of the original," +
+                            " made in leather with punched 3-Stripes," +
+                            " heel and tongue logos and lightweight step-in cushioning.",
+                    totalInventory = 5,
+                    variants = listOf(VariantItem("", "", "white/1", 0)),
+                    title = "Ultima show Running Shoes Pink",
+                    price = Price(
+                        amount = "172.00",
+                        currencyCode = "AED"
+                    ),
+                    discount = Discount(
+                        realPrice = "249.00",
+                        percent = 30
+                    ),
+                    vendor = "Adidas",
+                )
+            )
+        )
     }
 
-    override suspend fun getProductReviewById(productId: ID, reviewsCount: Int?): List<Review> {
-        TODO("Not yet implemented")
-    }
+    override suspend fun getProductReviewById(productId: ID, reviewsCount: Int?): List<Review> =
+        listOf(
+            Review(
+                reviewer = "",
+                rate = 2.5,
+                review = "Perfect phone",
+                description = "Its very nice experience Ga iloved it",
+                time = "2 months ago"
+            )
+        )
+
 
     override suspend fun setProductReview(productId: ID, review: Review) {
         TODO("Not yet implemented")
@@ -119,14 +176,14 @@ class FakeShopifyRepositoryImpl(
         emit(Resource.Success(products))
     }
 
-    override fun getProductsTag(): Flow<Resource<List<String>>> = flow {
+    override suspend fun getProductsTag(): Resource<List<String>> {
         val tags: List<String> = listOf("SHOES", "ACCESSORIES")
-        emit(Resource.Success(tags))
+        return Resource.Success(tags)
     }
 
-    override fun getProductsType(): Flow<Resource<List<String>>> = flow {
+    override suspend fun getProductsType(): Resource<List<String>>  {
         val types: List<String> = listOf("men", "women")
-        emit(Resource.Success(types))
+        return Resource.Success(types)
     }
 
     override suspend fun saveAddress(address: Storefront.MailingAddressInput): Resource<Boolean> {
@@ -165,8 +222,32 @@ class FakeShopifyRepositoryImpl(
         TODO("Not yet implemented")
     }
 
-    override fun getShopifyProductsByWishListIDs(): Flow<Resource<Product?>> {
-        TODO("Not yet implemented")
+    override fun getShopifyProductsByWishListIDs(): Flow<Resource<Product?>> = flow {
+        emit(
+            Resource.Success(
+                Product(
+                    images = listOf("https://www.skechers.com/dw/image/v2/BDCN_PRD/on/demandware.static/-/Sites-skechers-master/default/dw5fb9d39e/images/large/149710_MVE.jpg?sw=800"),
+                    description = "The Stan Smith owned the tennis court in the '70s." +
+                            " Today it runs the streets with the same clean," +
+                            " classic style." +
+                            " These kids' shoes preserve the iconic look of the original," +
+                            " made in leather with punched 3-Stripes," +
+                            " heel and tongue logos and lightweight step-in cushioning.",
+                    totalInventory = 5,
+                    variants = listOf(VariantItem("", "", "white/1", 0)),
+                    title = "Ultima show Running Shoes Pink",
+                    price = Price(
+                        amount = "172.00",
+                        currencyCode = "AED"
+                    ),
+                    discount = Discount(
+                        realPrice = "249.00",
+                        percent = 30
+                    ),
+                    vendor = "Adidas",
+                )
+            )
+        )
     }
 
     override suspend fun addProductWishListById(productId: ID) {
@@ -179,9 +260,23 @@ class FakeShopifyRepositoryImpl(
 
     override suspend fun getProductsByQuery(
         productQueryType: Constants.ProductQueryType,
-        queryContent: String
+        queryContent: String,
     ): Resource<Pageable<List<BrandProduct>>?> {
-        TODO("Not yet implemented")
+        return Resource.Success(
+            Pageable(
+                data = listOf(
+                    BrandProduct(
+                        id = ID("hamed"),
+                        title = "Ultima show Running Shoes Pink",
+                        images = listOf("https://www.skechers.com/dw/image/v2/BDCN_PRD/on/demandware.static/-/Sites-skechers-master/default/dw5fb9d39e/images/large/149710_MVE.jpg?sw=800"),
+                        isFavourite = wishList.contains(ID("hamed")),
+                        price = Storefront.MoneyV2().setAmount("200").setCurrencyCode(Storefront.CurrencyCode.SDG)
+                    )
+                ),
+                hasNext = true,
+                lastCursor = "Hamed"
+            )
+        )
     }
 
     override suspend fun getOrders(): Flow<Resource<List<Order>>> = flow {
@@ -270,5 +365,20 @@ class FakeShopifyRepositoryImpl(
     override suspend fun sendCompletePayment(): Resource<Pair<String?, String?>?> {
         TODO("Not yet implemented")
     }
+
+    override suspend fun changePassword(password: String): Resource<String?> {
+        TODO("Not yet implemented")
+    }
+
+    override suspend fun changePhoneNumber(phone: String): Resource<String?> {
+        TODO("Not yet implemented")
+    }
+
+    override suspend fun changeName(firstName: String, lastName: String): Resource<String?> {
+        TODO("Not yet implemented")
+    }
+
+    override suspend fun createUserEmail(email: String): Resource<Unit> =
+        Resource.Success(Unit)
 
 }
