+++ app/src/test/java/com/example/shopify/feature/navigation_bar/search/SearchViewModelTest.kt
@@ -1,14 +1,11 @@
 package com.example.shopify.feature.navigation_bar.search
 
-import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.example.shopify.MainDispatcherRule
-import com.example.shopify.data.repository.shopify.FakeShopifyRepository
+import com.example.shopify.feature.navigation_bar.model.repository.shopify.FakeShopifyRepositoryImpl
 import com.example.shopify.feature.navigation_bar.model.repository.shopify.ShopifyRepository
 import com.example.shopify.feature.search.view.SearchedProductsState
 import com.example.shopify.feature.search.viewModel.SearchViewModel
-import com.example.shopify.feature.wishList.viewModel.WishListViewModel
-import com.shopify.graphql.support.ID
 import junit.framework.Assert.assertEquals
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.delay
@@ -33,7 +30,7 @@ class SearchViewModelTest {
 
     @Before
     fun init() {
-        val repository: ShopifyRepository = FakeShopifyRepository()
+        val repository: ShopifyRepository = FakeShopifyRepositoryImpl()
         viewModel = SearchViewModel(
             repository = repository
         )
